+++ source/Energinet.DataHub.MarketParticipant.EntryPoint.WebApi/Controllers/ActorQueryController.cs
@@ -12,12 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using System.Threading.Tasks;
 using Energinet.DataHub.Core.App.Common.Abstractions.Users;
-using Energinet.DataHub.Core.App.Common.Security;
-using Energinet.DataHub.Core.App.WebApp.Authorization;
-using Energinet.DataHub.MarketParticipant.Application.Commands.Query.User;
+using Energinet.DataHub.MarketParticipant.Application.Commands.Query.Actor;
 using Energinet.DataHub.MarketParticipant.Application.Security;
 using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Extensions;
 using MediatR;
@@ -27,35 +24,37 @@ using Microsoft.Extensions.Logging;
 namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers;
 
 [ApiController]
-[Route("[controller]")]
-public sealed class UserOverviewController : ControllerBase
+[Route("query")]
+public sealed class ActorQueryController : ControllerBase
 {
-    private readonly ILogger<UserOverviewController> _logger;
+    private readonly ILogger<ActorContactController> _logger;
     private readonly IMediator _mediator;
     private readonly IUserContext<FrontendUser> _userContext;
 
-    public UserOverviewController(ILogger<UserOverviewController> logger, IMediator mediator, IUserContext<FrontendUser> userContext)
+    public ActorQueryController(
+        ILogger<ActorContactController> logger,
+        IMediator mediator,
+        IUserContext<FrontendUser> userContext)
     {
         _logger = logger;
         _mediator = mediator;
         _userContext = userContext;
     }
 
-    [HttpGet("users")]
-    [AuthorizeUser(Permission.UsersManage)]
-    public async Task<IActionResult> GetUserOverviewAsync(int pageNumber, int pageSize)
+    [HttpGet("selection-actors")]
+    public async Task<IActionResult> GetSelectionActorsAsync()
     {
         return await this.ProcessAsync(
             async () =>
-            {
-                var actorId = !_userContext.CurrentUser.IsFas
-                    ? _userContext.CurrentUser.ActorId
-                    : (Guid?)null;
+                {
+                    var command = new GetSelectionActorsQueryCommand(_userContext.CurrentUser.UserId);
 
-                var command = new GetUserOverviewCommand(pageNumber, pageSize, actorId);
-                var response = await _mediator.Send(command).ConfigureAwait(false);
-                return Ok(response.Users);
-            },
+                    var response = await _mediator
+                        .Send(command)
+                        .ConfigureAwait(false);
+
+                    return Ok(response.Actors);
+                },
             _logger).ConfigureAwait(false);
     }
 }