+++ source/Energinet.DataHub.MarketParticipant.IntegrationTests/Hosts/WebApi/GetSelectionActorsQueryHandlerIntegrationTests.cs
@@ -1,0 +1,61 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Linq;
+using System.Threading.Tasks;
+using Energinet.DataHub.Core.App.Common.Security;
+using Energinet.DataHub.MarketParticipant.Application.Commands.Query.Actor;
+using Energinet.DataHub.MarketParticipant.IntegrationTests.Common;
+using Energinet.DataHub.MarketParticipant.IntegrationTests.Fixtures;
+using MediatR;
+using Xunit;
+using Xunit.Categories;
+
+namespace Energinet.DataHub.MarketParticipant.IntegrationTests.Hosts.WebApi;
+
+[Collection("IntegrationTest")]
+[IntegrationTest]
+public sealed class GetSelectionActorsQueryHandlerIntegrationTests
+{
+    private readonly MarketParticipantDatabaseFixture _fixture;
+
+    public GetSelectionActorsQueryHandlerIntegrationTests(MarketParticipantDatabaseFixture fixture)
+    {
+        _fixture = fixture;
+    }
+
+    [Fact]
+    public async Task GetSelectionActors_GivenUserId_ReturnsActors()
+    {
+        // arrange
+        await using var host = await WebApiIntegrationTestHost.InitializeAsync(_fixture);
+        await using var scope = host.BeginScope();
+        var mediator = scope.GetInstance<IMediator>();
+
+        var (actorId, userId, externalUserId) = await _fixture.DatabaseManager.CreateUserAsync();
+
+        var roleTemplateId = await _fixture
+            .DatabaseManager
+            .AddUserPermissionsAsync(actorId, userId, new[] { Permission.UsersManage });
+
+        var command = new GetSelectionActorsQueryCommand(externalUserId);
+
+        // act
+        var actual = await mediator.Send(command);
+
+        // assert
+        Assert.Single(actual.Actors);
+        Assert.Equal(actorId, actual.Actors.Single().Id);
+    }
+}